Some under-the-hood corrections

+++ Ombuki/evaluation.py
@@ -2,10 +2,10 @@ from Ombuki.ombukiSolution import OmbukiSolution
 from problemInstance import ProblemInstance
 from typing import Tuple, List
 
-def ref_point(instance: ProblemInstance) -> Tuple[float, float]:
+def ref_point(instance: ProblemInstance) -> Tuple[float, int]:
     return 10000.0, instance.amount_of_vehicles + 1
 
-def calculate_Hypervolumes_area(nondominated_set: List[OmbukiSolution], ref_Hypervolumes: Tuple[float, float]) -> float:
+def calculate_Hypervolumes_area(nondominated_set: List[OmbukiSolution], ref_Hypervolumes: Tuple[float, int]) -> float:
     prev_TD, _ = ref_TD, ref_NV = ref_Hypervolumes[0], ref_Hypervolumes[1]
     area = 0.0
